+++ core/src/com/unciv/UncivGame.kt
@@ -94,6 +94,11 @@ open class UncivGame(val isConsoleMode: Boolean = false) : Game(), PlatformSpeci
         }
         Current = this
         files = UncivFiles(Gdx.files)
+        Concurrency.run {
+            // Delete temporary files created when downloading mods
+            val tempFiles = Gdx.files.local("mods").list().filter { !it.isDirectory && it.name().startsWith("temp-") }
+            for (file in tempFiles) file.delete()
+        }
 
         // If this takes too long players, especially with older phones, get ANR problems.
         // Whatever needs graphics needs to be done on the main thread,
@@ -194,7 +199,7 @@ open class UncivGame(val isConsoleMode: Boolean = false) : Game(), PlatformSpeci
             throw UncivShowableException("You are not allowed to spectate!")
         }
 
-        initializeResources(prevGameInfo, newGameInfo)
+        initializeResources(newGameInfo)
 
         val isLoadingSameGame = worldScreen != null && prevGameInfo != null && prevGameInfo.gameId == newGameInfo.gameId
         val worldScreenRestoreState = if (!callFromLoadScreen && isLoadingSameGame) worldScreen!!.getRestoreState() else null
@@ -233,16 +238,12 @@ open class UncivGame(val isConsoleMode: Boolean = false) : Game(), PlatformSpeci
     }
 
     /** The new game info may have different mods or rulesets, which may use different resources that need to be loaded. */
-    private suspend fun initializeResources(prevGameInfo: GameInfo?, newGameInfo: GameInfo) {
-        if (prevGameInfo == null
-                || prevGameInfo.gameParameters.baseRuleset != newGameInfo.gameParameters.baseRuleset
-                || prevGameInfo.gameParameters.mods != newGameInfo.gameParameters.mods) {
-            withGLContext {
-                ImageGetter.setNewRuleset(newGameInfo.ruleset)
-            }
-            val fullModList = newGameInfo.gameParameters.getModsAndBaseRuleset()
-            musicController.setModList(fullModList)
+    private suspend fun initializeResources(newGameInfo: GameInfo) {
+        withGLContext {
+            ImageGetter.setNewRuleset(newGameInfo.ruleset)
         }
+        val fullModList = newGameInfo.gameParameters.getModsAndBaseRuleset()
+        musicController.setModList(fullModList)
     }
 
     /** Re-creates the current [worldScreen], if there is any. */
@@ -480,7 +481,7 @@ open class UncivGame(val isConsoleMode: Boolean = false) : Game(), PlatformSpeci
 
     companion object {
         //region AUTOMATICALLY GENERATED VERSION DATA - DO NOT CHANGE THIS REGION, INCLUDING THIS COMMENT
-        val VERSION = Version("4.12.8", 1019)
+        val VERSION = Version("4.12.9", 1020)
         //endregion
 
         /** Global reference to the one Gdx.Game instance created by the platform launchers - do not use without checking [isCurrentInitialized] first. */