+++ core/src/com/unciv/logic/automation/civilization/UseGoldAutomation.kt
@@ -5,13 +5,11 @@ import com.unciv.logic.civilization.Civilization
 import com.unciv.logic.map.BFS
 import com.unciv.logic.map.tile.Tile
 import com.unciv.models.ruleset.INonPerpetualConstruction
-import com.unciv.models.ruleset.PerpetualConstruction
 import com.unciv.models.ruleset.Victory
 import com.unciv.models.ruleset.tile.ResourceType
 import com.unciv.models.ruleset.unique.UniqueType
 import com.unciv.models.stats.Stat
-import java.util.SortedMap
-import java.util.TreeMap
+import java.util.*
 
 object UseGoldAutomation {
 
@@ -23,11 +21,11 @@ object UseGoldAutomation {
 
         for (city in civ.cities.sortedByDescending { it.population.population }) {
             val construction = city.cityConstructions.getCurrentConstruction()
-            if (construction is PerpetualConstruction) continue
-            if ((construction as INonPerpetualConstruction).canBePurchasedWithStat(city, Stat.Gold)
-                && city.civ.gold / 3 >= construction.getStatBuyCost(city, Stat.Gold)!!) {
-                city.cityConstructions.purchaseConstruction(construction, 0, true)
-            }
+            if (construction !is INonPerpetualConstruction) continue
+            val statBuyCost = construction.getStatBuyCost(city, Stat.Gold) ?: continue
+            if (!city.cityConstructions.isConstructionPurchaseAllowed(construction, Stat.Gold, statBuyCost)) continue
+            if (civ.gold < statBuyCost / 3) continue
+            city.cityConstructions.purchaseConstruction(construction, 0, true)
         }
 
         maybeBuyCityTiles(civ)