+++ core/src/com/unciv/logic/civilization/Civilization.kt
@@ -887,10 +887,6 @@ class Civilization : IsPartOfGameInfoSerialization {
 
     fun addNotification(text: String, actions: Iterable<NotificationAction>?, category: NotificationCategory, vararg notificationIcons: String) {
         if (playerType == PlayerType.AI) return // no point in lengthening the saved game info if no one will read it
-        if (notifications.lastOrNull()?.let { it.text == text && it.category == category && it.icons == notificationIcons.toList() } == true) {
-            Log.debug("Duplicate notification \"%s\"", text)
-            return
-        }
         notifications.add(Notification(text, notificationIcons, actions, category))
     }
     // endregion