+++ core/src/com/unciv/ui/components/widgets/TranslatedSelectBox.kt
@@ -1,11 +1,11 @@
 package com.unciv.ui.components.widgets
 
 import com.badlogic.gdx.scenes.scene2d.ui.SelectBox
-import com.badlogic.gdx.scenes.scene2d.ui.Skin
 import com.badlogic.gdx.utils.Array
 import com.unciv.models.translations.tr
+import com.unciv.ui.screens.basescreen.BaseScreen
 
-class TranslatedSelectBox(values: Collection<String>, default: String, skin: Skin) : SelectBox<TranslatedSelectBox.TranslatedString>(skin) {
+class TranslatedSelectBox(values: Collection<String>, default: String) : SelectBox<TranslatedSelectBox.TranslatedString>(BaseScreen.skin) {
     class TranslatedString(val value: String) {
         val translation = value.tr(hideIcons = true)
         override fun toString() = translation