+++ core/src/com/unciv/ui/screens/newgamescreen/MapParametersTable.kt
@@ -122,7 +122,7 @@ class MapParametersTable(
             add(optionsTable).colspan(2).grow().row()
         } else {
             val mapShapeSelectBox =
-                    TranslatedSelectBox(mapShapes, mapParameters.shape, skin)
+                    TranslatedSelectBox(mapShapes, mapParameters.shape)
             mapShapeSelectBox.onChange {
                 mapParameters.shape = mapShapeSelectBox.selected.value
                 updateWorldSizeTable()
@@ -161,7 +161,7 @@ class MapParametersTable(
             }
             add(optionsTable).colspan(2).grow().row()
         } else {
-            mapTypeSelectBox = TranslatedSelectBox(mapTypes, mapParameters.type, skin)
+            mapTypeSelectBox = TranslatedSelectBox(mapTypes, mapParameters.type)
 
             mapTypeSelectBox.onChange {
                 mapParameters.type = mapTypeSelectBox.selected.value
@@ -190,7 +190,7 @@ class MapParametersTable(
             add(optionsTable).colspan(2).grow().row()
         } else {
             val mapSizes = MapSize.names() + listOf(MapSize.custom)
-            worldSizeSelectBox = TranslatedSelectBox(mapSizes, mapParameters.mapSize.name, skin)
+            worldSizeSelectBox = TranslatedSelectBox(mapSizes, mapParameters.mapSize.name)
             worldSizeSelectBox.onChange { updateWorldSizeTable() }
 
             addHexagonalSizeTable()
@@ -272,7 +272,7 @@ class MapParametersTable(
             }
             add(optionsTable).colspan(2).grow().row()
         } else {
-            resourceSelectBox = TranslatedSelectBox(mapResources, mapParameters.mapResources, skin)
+            resourceSelectBox = TranslatedSelectBox(mapResources, mapParameters.mapResources)
 
             resourceSelectBox.onChange {
                 mapParameters.mapResources = resourceSelectBox.selected.value